requirements.txt
blinker==1.3
Flask==0.10.1
Flask-DebugToolbar==0.10.0
Flask-SQLAlchemy==2.0
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
nltk==3.2.1
oauthlib==1.1.1
psycopg2==2.6.1
python-twitter==2.2
pytz==2014.2
requests==2.10.0
requests-oauthlib==0.6.1
simplejson==3.4.1
SQLAlchemy==1.0.3
Werkzeug==0.10.4
yahoo-finance==1.2.1
